requirements_dev.txt
-r ./requirements.txt
ipython==5.3.0
ipdb==0.10.3
ipdbplugin==1.4.5
django-debug-toolbar==1.7
django-nose==1.4.4
nose-exclude==0.5.0
mock==2.0.0
memory-profiler==0.45
flake8==3.3.0
ptpython==0.33
ptpdb==0.17
remote-pdb==1.2.0
